requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
Flask-SQLAlchemy==3.1.1
SQLAlchemy==2.0.23
openpyxl==3.1.2
python-dateutil==2.8.2
Werkzeug==3.0.0
Pillow>=10.1.0
